LayoutTests/imported/w3c/web-platform-tests/css/css-grid/layout-algorithm/baseline-alignment-affects-intrinsic-size-001-expected.txt

Horizontal 4x1 grid with parallel and orthogonal items.

É
É
É
É
É
É
É
É
É
É
É
É
É
É
É
É


FAIL #grid1-auto-auto 1 assert_equals:
<div id="grid1-auto-auto" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(auto, auto);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-auto-auto 2 assert_equals:
<div id="grid2-auto-auto" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(auto, auto);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-auto-auto 3
PASS #grid4-auto-auto 4
FAIL #grid1-auto-min-content 5 assert_equals:
<div id="grid1-auto-min-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(auto, min-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-auto-min-content 6 assert_equals:
<div id="grid2-auto-min-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(auto, min-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-auto-min-content 7
PASS #grid4-auto-min-content 8
FAIL #grid1-auto-max-content 9 assert_equals:
<div id="grid1-auto-max-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(auto, max-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-auto-max-content 10 assert_equals:
<div id="grid2-auto-max-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(auto, max-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-auto-max-content 11
PASS #grid4-auto-max-content 12
FAIL #grid1-min-content-auto 13 assert_equals:
<div id="grid1-min-content-auto" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(min-content, auto);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-min-content-auto 14 assert_equals:
<div id="grid2-min-content-auto" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(min-content, auto);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-min-content-auto 15
PASS #grid4-min-content-auto 16
FAIL #grid1-min-content-min-content 17 assert_equals:
<div id="grid1-min-content-min-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(min-content, min-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-min-content-min-content 18 assert_equals:
<div id="grid2-min-content-min-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(min-content, min-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-min-content-min-content 19
PASS #grid4-min-content-min-content 20
FAIL #grid1-min-content-max-content 21 assert_equals:
<div id="grid1-min-content-max-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(min-content, max-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-min-content-max-content 22 assert_equals:
<div id="grid2-min-content-max-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(min-content, max-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-min-content-max-content 23
PASS #grid4-min-content-max-content 24
FAIL #grid1-max-content-auto 25 assert_equals:
<div id="grid1-max-content-auto" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(max-content, auto);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-max-content-auto 26 assert_equals:
<div id="grid2-max-content-auto" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(max-content, auto);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-max-content-auto 27
PASS #grid4-max-content-auto 28
FAIL #grid1-max-content-min-content 29 assert_equals:
<div id="grid1-max-content-min-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(max-content, min-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-max-content-min-content 30 assert_equals:
<div id="grid2-max-content-min-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(max-content, min-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-max-content-min-content 31
PASS #grid4-max-content-min-content 32
FAIL #grid1-max-content-max-content 33 assert_equals:
<div id="grid1-max-content-max-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(max-content, max-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-max-content-max-content 34 assert_equals:
<div id="grid2-max-content-max-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(max-content, max-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-max-content-max-content 35
PASS #grid4-max-content-max-content 36
FAIL #grid1-0px-auto 37 assert_equals:
<div id="grid1-0px-auto" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(0px, auto);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-0px-auto 38 assert_equals:
<div id="grid2-0px-auto" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(0px, auto);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-0px-auto 39
PASS #grid4-0px-auto 40
FAIL #grid1-0px-min-content 41 assert_equals:
<div id="grid1-0px-min-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(0px, min-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-0px-min-content 42 assert_equals:
<div id="grid2-0px-min-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(0px, min-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-0px-min-content 43
PASS #grid4-0px-min-content 44
FAIL #grid1-0px-max-content 45 assert_equals:
<div id="grid1-0px-max-content" class="grid contentStart itemsBaseline" data-expected-width="112" style="grid-template-columns: minmax(0px, max-content);">
    <div class="item" data-offset-x="36" data-offset-y="10" data-expected-width="40" data-expected-height="38">É</div>
    <div class="item verticalLR" data-offset-x="18" data-offset-y="62" data-expected-width="50" data-expected-height="48">É</div>
    <div class="item" data-offset-x="36" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="12" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 112 but got 98
FAIL #grid2-0px-max-content 46 assert_equals:
<div id="grid2-0px-max-content" class="grid contentStart itemsBaseline" data-expected-width="165" style="grid-template-columns: minmax(0px, max-content);">
    <div class="item extraRightPadding" data-offset-x="72" data-offset-y="10" data-expected-width="87" data-expected-height="38">É</div>
    <div class="item verticalLR extraLeftPadding" data-offset-x="12" data-offset-y="62" data-expected-width="92" data-expected-height="48">É</div>
    <div class="item" data-offset-x="72" data-offset-y="124" data-expected-width="70" data-expected-height="68">É</div>
    <div class="item verticalLR" data-offset-x="48" data-offset-y="206" data-expected-width="80" data-expected-height="78">É</div>
</div>
width expected 165 but got 134
PASS #grid3-0px-max-content 47
PASS #grid4-0px-max-content 48
